refactor: extract shared api surface into ferriskey-api-core
