+++ api/src/application/http/server/app_state.rs
@@ -1,17 +1,1 @@
-use std::sync::Arc;
-
-use ferriskey_core::ApplicationService;
-
-use crate::args::Args;
-
-#[derive(Clone)]
-pub struct AppState {
-    pub args: Arc<Args>,
-    pub service: ApplicationService,
-}
-
-impl AppState {
-    pub fn new(args: Arc<Args>, service: ApplicationService) -> Self {
-        Self { args, service }
-    }
-}
+pub use ferriskey_api_core::app_state::*;
