+++ libs/ferriskey-api-core/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "ferriskey-api-core"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+ferriskey-core = { path = "../../core" }
+anyhow = "1.0.97"
+axum = "0.8.8"
+axum-cookie = "0.2.3"
+axum-extra = { version = "0.12.5", features = ["cookie", "typed-header"] }
+axum-macros = "0.5.0"
+base64 = "0.22.1"
+base32 = "0.5.1"
+chrono = { version = "0.4.40", features = ["serde"] }
+clap = { version = "4.5.32", features = ["derive", "env"] }
+regex = "1.11.1"
+serde = "1.0.228"
+serde_json = "1.0.148"
+thiserror = "2.0.17"
+tracing = "0.1.44"
+utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
+uuid = { version = "1.19.0", features = ["serde", "v4", "v7"] }
+validator = { version = "0.20.0", features = ["derive"] }
+url = "2.5.4"
+urlencoding = "2.1.3"
+webauthn-rs = "0.5.2"
