+++ libs/ferriskey-api-core/src/api_entities.rs
@@ -1,0 +1,4 @@
+pub mod api_error;
+pub mod api_response_body;
+pub mod api_success;
+pub mod response;
