+++ libs/ferriskey-api-core/src/api_entities/api_error.rs
