+++ libs/ferriskey-api-core/src/api_entities/response.rs
