+++ libs/ferriskey-api-core/src/app_state.rs
@@ -1,0 +1,17 @@
+use std::sync::Arc;
+
+use ferriskey_core::ApplicationService;
+
+use crate::args::Args;
+
+#[derive(Clone)]
+pub struct AppState {
+    pub args: Arc<Args>,
+    pub service: ApplicationService,
+}
+
+impl AppState {
+    pub fn new(args: Arc<Args>, service: ApplicationService) -> Self {
+        Self { args, service }
+    }
+}
